# Strip environment from exported function