@odelmarcelle 2022-03-14
After some testing, it does not work as expected because `...future.FUN` is still evaluated with it's own environment. So it doesn't seem to make used of exported globals. I will look a bit for a way to solve that.