@odelmarcelle 2022-03-15
I don't see how to solve the issue safely for all possible usage of `future.apply`. A satisfactory solution would probably require numerous changes to `future.apply` and the cluster class of `future`.

In the absence of an easy-to-implement solution, I'm closing this pull request. If `future.apply` is not capable of efficiently stripping the unwanted function environment, it is up to the user to be careful about the environment of the provided function.